Move pickup style to Item Pool tab
This makes these available for Prime 1
