+++ randovania/gui/ui_files/preset_item_pool.ui
@@ -113,11 +113,83 @@
          </widget>
         </item>
         <item>
+         <widget class="QGroupBox" name="pickup_style_box">
+          <property name="sizePolicy">
+           <sizepolicy hsizetype="Preferred" vsizetype="Fixed">
+            <horstretch>0</horstretch>
+            <verstretch>0</verstretch>
+           </sizepolicy>
+          </property>
+          <property name="title">
+           <string>Pickup style</string>
+          </property>
+          <layout class="QVBoxLayout" name="verticalLayout_2">
+           <item>
+            <widget class="QComboBox" name="pickup_model_combo">
+             <item>
+              <property name="text">
+               <string>Use correct item model, scan and name</string>
+              </property>
+             </item>
+             <item>
+              <property name="text">
+               <string>Use correct scan and name, hide the model</string>
+              </property>
+             </item>
+             <item>
+              <property name="text">
+               <string>Use correct name, hide the model and scan</string>
+              </property>
+             </item>
+             <item>
+              <property name="text">
+               <string>Hide the model, scan and name</string>
+              </property>
+             </item>
+            </widget>
+           </item>
+           <item>
+            <widget class="QLabel" name="pickup_data_source_label">
+             <property name="text">
+              <string>When hiding some part of the pickup, it's replaced with:</string>
+             </property>
+            </widget>
+           </item>
+           <item>
+            <widget class="QComboBox" name="pickup_data_source_combo">
+             <item>
+              <property name="text">
+               <string>Energy Transfer Module/Nothing's data</string>
+              </property>
+             </item>
+             <item>
+              <property name="text">
+               <string>A random item data</string>
+              </property>
+             </item>
+             <item>
+              <property name="text">
+               <string>The data of the pickup in that place</string>
+              </property>
+             </item>
+            </widget>
+           </item>
+          </layout>
+         </widget>
+        </item>
+        <item>
          <widget class="QGroupBox" name="random_starting_box">
           <property name="title">
            <string>Random Starting Items</string>
           </property>
           <layout class="QGridLayout" name="gridLayout_2">
+           <item row="2" column="1">
+            <widget class="QSpinBox" name="maximum_starting_spinbox">
+             <property name="maximum">
+              <number>30</number>
+             </property>
+            </widget>
+           </item>
            <item row="2" column="0">
             <widget class="QLabel" name="maximum_starting_label">
              <property name="text">
@@ -135,8 +207,8 @@
              </property>
             </widget>
            </item>
-           <item row="2" column="1">
-            <widget class="QSpinBox" name="maximum_starting_spinbox">
+           <item row="1" column="1">
+            <widget class="QSpinBox" name="minimum_starting_spinbox">
              <property name="maximum">
               <number>30</number>
              </property>
@@ -149,13 +221,6 @@
              </property>
              <property name="alignment">
               <set>Qt::AlignLeading|Qt::AlignLeft|Qt::AlignVCenter</set>
-             </property>
-            </widget>
-           </item>
-           <item row="1" column="1">
-            <widget class="QSpinBox" name="minimum_starting_spinbox">
-             <property name="maximum">
-              <number>30</number>
              </property>
             </widget>
            </item>
